+++ x/auth/vesting/types/test_common.go
@@ -1,0 +1,29 @@
+// nolint noalias
+package types
+
+import (
+	"github.com/tendermint/tendermint/crypto"
+	"github.com/tendermint/tendermint/crypto/secp256k1"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+)
+
+// NewTestMsg generates a test message
+func NewTestMsg(addrs ...sdk.AccAddress) *sdk.TestMsg {
+	return sdk.NewTestMsg(addrs...)
+}
+
+// NewTestCoins coins to more than cover the fee
+func NewTestCoins() sdk.Coins {
+	return sdk.Coins{
+		sdk.NewInt64Coin("atom", 10000000),
+	}
+}
+
+// KeyTestPubAddr generates a test key pair
+func KeyTestPubAddr() (crypto.PrivKey, crypto.PubKey, sdk.AccAddress) {
+	key := secp256k1.GenPrivKey()
+	pub := key.PubKey()
+	addr := sdk.AccAddress(pub.Address())
+	return key, pub, addr
+}